Factored out all the metadata calculations into a separate contract...
We were hitting contract size limit when trying to do everything in the
`Staker`. To generate metadata on-chain will require a separate contract.

This is akin to the `NonfungibleTokenPositionDescriptor` in UniswapV3.

Updated tests to reflect new deployment structure of `Staker`, which
takes a `PositionMetadata` contract address as its only argument now.
Previously, it had no arguments to the constructor.